requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
pip==22.0.4
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==58.1.0
urllib3==2.3.0